requirements.txt
streamlit
pillow
tensorflow 
opencv-python 
numpy
lxml[html_clean]
opencv-python-headless